refactor(shared): consolidate CLI helpers under internal/shared
Move prompt, IAM display, and federated identity parsing helpers into a
single importable package and drop the duplicated cmd/iam/internal and
kubernetes/iam/shared wrappers.
