+++ internal/shared/federated.go
